requirements.txt
alpaca-py
numpy
pandas
python-dotenv
pytest
yfinance
